+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2025-12-01
+
+### Added
+- Extension methods for energy consumption calculations
+
+### Changed
+- Refactored services.dart for better code organization
+
+[0.2.0]: https://github.com/nvf-research/energeist/compare/v0.1.0...v0.2.0
+
 ## [0.1.0] - 2024-11-29
 
 ### Added